requirements.txt
pytest==8.4.2
python-dotenv==1.1.1
selenium==4.35.0
webdriver-manager==4.0.2
